refactor file content
