+++ main.py
@@ -1,8 +1,9 @@
 import os
 import matplotlib.pyplot as plt
 from environment import StochasticGameStartEnv
+from pipeline import LoggingStudy
 
-from pipeline.study_agent import LoggingStudy, load_dqn_agent
+from pipeline.study_agent import load_dqn_agent
 from algorithms.deep_q_networks import DQN, DoubleDQN
 from environment.environment import ProductOwnerEnv, CreditPayerEnv
