+++ pipeline/__init__.py
@@ -1,1 +1,3 @@
+from .base_study import BaseStudy
 from .metrics_study import MetricsStudy
+from .logging_study import LoggingStudy
